backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pymongo
python-dotenv
spacy
openai
pdfplumber
python-docx
reportlab
